# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Our Recipes API
  version: 0.1.0
  description: Local household foundation API. Profiles are not authentication.
  license:
    name: Private household use — see repository license decision before redistribution
servers:
  - url: /api/v1
security: []
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      summary: List household profiles
      responses:
        '200': { description: Profile list }
        '503': { description: Storage is unavailable }
    post:
      operationId: addProfile
      summary: Add a household profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileInput' }
      responses:
        '201': { description: Profile created }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { $ref: '#/components/responses/UntrustedOrigin' }
...
  /tags/{tagName}:
    parameters:
      - name: tagName
        in: path
        required: true
        schema: { type: string, maxLength: 40 }
    patch:
      operationId: updateTag
      summary: Rename or recolor a tag while preserving recipe associations
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagInput' }
      responses:
        '200': { description: Tag updated }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { $ref: '#/components/responses/UntrustedOrigin' }
        '404': { description: Tag not found }
        '409': { description: Target tag exists; use merge }
    delete:
      operationId: deleteTag
      summary: Remove a tag association from every recipe
      responses:
        '204': { description: Tag removed }
        '403': { $ref: '#/components/responses/UntrustedOrigin' }
        '404': { description: Tag not found }
...
  /recipes:
    get:
      operationId: listRecipes
      summary: Search a paginated, lifecycle-aware shared recipe library
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: creator
          in: query
          schema: { type: string, format: uuid }
        - name: tag
          in: query
          schema: { type: string, maxLength: 40 }
        - name: collection
          in: query
          schema: { type: string, format: uuid }
        - name: category
          in: query
          schema: { type: string, maxLength: 80 }
        - name: cuisine
          in: query
          schema: { type: string, maxLength: 80 }
        - name: maxTotalMinutes
          in: query
          schema: { type: integer, minimum: 1, maximum: 30240 }
        - name: favorite
          in: query
          schema: { type: boolean }
        - name: cooked
          in: query
          schema: { type: boolean }
        - name: status
          in: query
          schema: { type: string, enum: [active, archived, trash, all], default: active }
        - name: sort
          in: query
          schema:
            {
              type: string,
              enum:
                [
                  recently-added,
                  recently-updated,
                  alphabetical,
                  most-recently-cooked,
                  shortest-time,
                ],
            }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
      responses:
        '200': { description: Recipe cards }
        '400': { description: Invalid search query }
    post:
      operationId: createRecipe
      summary: Create a structured recipe and first revision
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecipeInput' }
      responses:
        '201': { description: Recipe created }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { $ref: '#/components/responses/UntrustedOrigin' }
        '409': { description: Profile selection required }
...
  /recipes/{recipeId}/images:
    parameters:
      - name: recipeId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: createRecipeImage
      summary: Add a bounded local recipe photo and return normalized media metadata
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                  description: JPEG, PNG, or WebP source image, 10 MB maximum.
                altText:
                  type: string
                  maxLength: 180
      responses:
        '201': { description: Normalized photo metadata created }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { $ref: '#/components/responses/UntrustedOrigin' }
        '404': { description: Recipe not found }
        '409': { description: Profile selection required }
        '413': { description: Image exceeds the byte limit }
  /recipes/{recipeId}/images/{imageId}:
    parameters:
      - name: recipeId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: imageId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getRecipeImage
      summary: Read a normalized local WebP recipe photo
      responses:
        '200':
          description: Normalized image bytes
          content:
            image/webp:
              schema: { type: string, format: binary }
        '404': { description: Recipe photo not found }
    delete:
      operationId: deleteRecipeImage
      summary: Remove a local recipe photo and its media metadata
      responses:
        '204': { description: Recipe photo removed }
        '403': { $ref: '#/components/responses/UntrustedOrigin' }
        '404': { description: Recipe photo not found }
        '409': { description: Profile selection required }
...
  /meal-plan:
    get:
      operationId: listMealPlan
      summary: Read planned meals for an inclusive date range
      responses:
        '200': { description: Planned meals }
        '400': { description: Invalid week range }
    post:
      operationId: addMealPlanEntry
      summary: Add an existing recipe to a date and meal
      responses:
        '201': { description: Planned meal created }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { $ref: '#/components/responses/UntrustedOrigin' }
...
components:
  schemas:
    ProfileInput:
      type: object
      required: [displayName, color, units, temperatureUnit, locale, timezone]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 80 }
        color: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
        avatarUrl: { type: string, format: uri }
        units: { type: string, enum: [metric, imperial] }
        temperatureUnit: { type: string, enum: [C, F] }
        locale: { type: string }
        timezone: { type: string }
...
    TagInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
        color: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
...
    RecipeInput:
      type: object
      required:
        [title, servings, prepMinutes, cookMinutes, tags, ingredientGroups, instructionSections]
      properties:
        title: { type: string, minLength: 1, maxLength: 160 }
        summary: { type: string, maxLength: 800 }
        status: { type: string, enum: [active, archived, trash], default: active }
        servings: { type: string, minLength: 1, maxLength: 80 }
        prepMinutes: { type: integer, minimum: 0 }
        cookMinutes: { type: integer, minimum: 0 }
        restMinutes: { type: integer, minimum: 0, default: 0 }
        difficulty: { type: string, maxLength: 40 }
        cuisine: { type: string, maxLength: 80 }
        category: { type: string, maxLength: 80 }
        tips: { type: string, maxLength: 2000 }
        sharedNotes: { type: string, maxLength: 2000 }
        sourceName: { type: string }
        sourceUrl: { type: string, format: uri }
        originalAuthor: { type: string, maxLength: 160 }
        cookingMethod: { type: string, maxLength: 80 }
        equipment:
          type: array
          maxItems: 30
          items: { type: string, maxLength: 120 }
        nutritionCalories: { type: number, minimum: 0 }
        nutritionProteinGrams: { type: number, minimum: 0 }
        nutritionCarbohydrateGrams: { type: number, minimum: 0 }
        nutritionFatGrams: { type: number, minimum: 0 }
        nutritionFiberGrams: { type: number, minimum: 0 }
        tags: { type: array, items: { type: string } }
        ingredientGroups: { type: array, items: { type: object } }
        instructionSections: { type: array, items: { type: object } }
...
  responses:
    InvalidRequest: { description: Input did not meet the contract }
    UntrustedOrigin: { description: The browser origin was not explicitly trusted }